QAC_practice_Learn/qc_tweet_proof_demo/requirements.txt
qiskit
matplotlib
pillow
pandas
